refactor(project): rename project packages
Rename project packages for better clarity and organization.
