+++ src/main/java/com/steve/graphic/Material.java
@@ -17,7 +17,7 @@
 
 import org.slf4j.Logger;
 
-import com.steve.utils.LogUtil;
+import com.steve.util.LogUtil;
 
 public class Material {
