+++ src/main/java/com/steve/graphic/Texture.java
@@ -41,9 +41,10 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 
 import org.slf4j.Logger;
-import org.apache.commons.lang3.StringUtils;
 
-import com.steve.utils.LogUtil;
+import com.steve.util.LogUtil;
+
+import org.apache.commons.lang3.StringUtils;
 
 /**
  * The {@code Texture} class is responsible for loading and managing
