+++ src/main/resources/model/bag.obj
