+++ src/main/resources/model/cube.mtl
