+++ src/main/resources/model/cube.obj
