+++ src/main/resources/model/monkey.obj
