+++ src/main/resources/model/nile.mtl
