+++ src/main/resources/model/objTest.mtl
