+++ src/main/resources/model/objTest.obj
