+++ src/main/resources/model/ttqobj.mtl
