+++ src/main/resources/model/ttqobj.obj
